Get all genes

Return all known genes from a given species.

all_genes(
  species,
  method = c("gprofiler", "homologene", "babelgene"),
  ensure_filter_nas = FALSE,
  run_map_species = TRUE,
  force = FALSE,
  verbose = TRUE,
  ...
)

Arguments

species

Species to get all genes for. Will first be standardised with map_species.

method

R package to use for gene mapping:

"gprofiler"

Slower but more species and genes.

"homologene"

Faster but fewer species and genes.

"babelgene"

Faster but fewer species and genes. Also gives consensus scores for each gene mapping based on a several different data sources.

ensure_filter_nas

Perform an extra check to remove genes that are NAs of any kind.

run_map_species

Standardise species names with map_species first (Default: TRUE).

force

Force a new files to be generated even if a cached one exists.

verbose

Print messages.

...

Additional arguments to be passed to gorth or homologene.

NOTE: To return only the most "popular" interspecies ortholog mappings, supply mthreshold=1 here AND set method="gprofiler" above. This procedure tends to yield a greater number of returned genes but at the cost of many of them not being true biological 1:1 orthologs.

For more details, please see here.

Value

Table with all gene symbols from the given species.

Details

References homologeneData or gconvert.

Examples

genome_mouse <- all_genes(species = "mouse")
#> Retrieving all genes using: gprofiler
#> Retrieving all organisms available in gprofiler.
#> Using stored `gprofiler_orgs`.
#> Mapping species name: mouse
#> Common name mapping found for mouse
#> 1 organism identified from search: mmusculus
#> Using cached file: /github/home/.cache/R/orthogene/all_genes-mmusculus-gprofiler.csv.gz
#> Returning all 78,202 genes from mouse.
genome_human <- all_genes(species = "human")
#> Retrieving all genes using: gprofiler
#> Retrieving all organisms available in gprofiler.
#> Using stored `gprofiler_orgs`.
#> Mapping species name: human
#> Common name mapping found for human
#> 1 organism identified from search: hsapiens
#> Using cached file: /github/home/.cache/R/orthogene/all_genes-hsapiens-gprofiler.csv.gz
#> Returning all 78,687 genes from human.