Map genes from one species to another

Currently supports 700+ species. See map_species() for a full list of available species.

convert_orthologs(
  gene_df,
  gene_input = "rownames",
  gene_output = "rownames",
  standardise_genes = FALSE,
  input_species,
  output_species = "human",
  method = c("gprofiler", "homologene"),
  drop_nonorths = TRUE,
  non121_strategy = "drop_both_species",
  as_sparse = FALSE,
  sort_rows = FALSE,
  verbose = TRUE,
  ...
)

Arguments

gene_df
Data object containing the genes (see gene_input for options on how the genes can be stored within the object).

Can be one of the following formats:

matrixA sparse or dense matrix.

data.frameA data.frame, data.table. or tibble.

codelistA list or vector.

Genes can be in any format (e.g. HGNC symbols, ENSEMBL IDs, UCSC) and will be automatically converted to HGNC gene symbols unless specified otherwise with the ... arguments.

gene_input
Which aspect of gene_df you want to get gene names from:

"rownames" : From row names of data.frame/matrix.

"colnames" : From column names of data.frame/matrix.

<column name> : From a column in gene_df, e.g. "gene_names".

gene_output
How to return genes. Options include:

"rownames" : As row names of gene_df.

"colnames" : As column names of gene_df.

"columns" : As new columns "input_gene", "ortholog_gene" (and "input_gene_standard" if standardise_genes=TRUE) in gene_df.

"dict" : As a dictionary (named list) where the names are input_gene and the values are ortholog_gene.

"dict_rev" : As a reversed dictionary (named list) where the names are ortholog_gene and the values are input_gene.

standardise_genes
If TRUE genes will first be standardised to HGNC symbols using gconvert. This will add a new column to gene_df called "input_gene_standard".

input_species
Name of the input species (e.g., "mouse","fly"). See taxa_id_dict() for a full list of available species.

output_species
Name of the output species (e.g. "human").

method
R package to to use for gene mapping: "gprofiler" (slower but more species and genes) or "homologene" (faster but fewer species and genes).

drop_nonorths
Drop genes that don't have an ortholog in the output_species.

non121_strategy
How to handle genes that don't have 1:1 mappings between input_species:output_species. Options include:

"drop_both_species" or "dbs" or 1 : Drop genes that have duplicate mappings in either the input_species or output_species (Default).

"drop_input_species" or "dis" or 2 : Drop genes that have duplicate mappings in the input_species.

"drop_output_species" or "dos" or 3 : Drop genes that have duplicate mappings in the output_species.

"keep_both_species" or "kbs" or 4 : Keep all genes regardless of whether they have duplicate mappings in either species.

as_sparse
Convert gene_df to a sparse matrix. Only works if gene_df is one of the following classes:

matrix

Matrix

data.frame

data.table

tibble

If gene_df is a sparse matrix to begin with, it will be returned as a sparse matrix (so long as gene_output= "rownames" or "colnames").

sort_rows
Sort gene_df rows alphanumerically.

verbose
Print messages.

...
Additional arguments to be passed to gconvert or homologene.

Value

gene_df with orthologs converted to the output_species. Instead returned as a dictionary (named list) if gene_output="dict" or "dict_rev".

Examples

data("exp_mouse")
gene_df <- convert_orthologs(gene_df = exp_mouse,
                             input_species="mouse")

#> WARNING: In order to set gene_output='rownames' must set non121_strategy='drop_both_species'.
#>  Setting non121_strategy='drop_both_species'.

#> Preparing gene_df.

#> Extracting genes from rownames.

#> Converting mouse ==> human orthologs using: gprofiler

#> Using stored `gprofiler_orgs`.

#> Mapping species name: mouse

#> Common name mapping found for mouse

#> 1 organism identified from search: mmusculus

#> Using stored `gprofiler_orgs`.

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: hsapiens

#> Checking for genes without orthologs in human.

#> Extracting genes from input_gene.

#> Extracting genes from ortholog_gene.

#> + Dropping 2,769 NAs of all kinds from ortholog_gene.

#> Checking for genes without 1:1 orthologs.

#> + Dropping 478 genes that have multiple input_gene per ortholog_gene.

#> + Dropping 299 genes that have multiple ortholog_gene per input_gene.

#> Setting ortholog_gene to rownames.

#> 
#> =========== REPORT SUMMARY ===========

#> Total genes dropped during inter-species conversion:  3,025 / 15,259 (20%)