Identify the number of orthologous genes between two species

ewceData::mouse_to_human_homologs() provides 15,604 1:1 mouse:human orthologs. homologene-based functions like EWCE::convert_orthologs provide an improved 16,766 1:1 mouse:human orthologs.

report_orthologs(
  target_species = "mouse",
  reference_species = "human",
  standardise_genes = FALSE,
  method = c("homologene", "gorth"),
  drop_nonorths = TRUE,
  one_to_one_only = TRUE,
  round_digits = 2,
  return_report = TRUE,
  verbose = TRUE
)

Arguments

target_species
Target species.

reference_species
Reference species.

standardise_genes
If TRUE genes will first be standardised to HGNC symbols using gprofiler2::gconvert. This will add a new column to gene_df called "input_gene_standard".

method
R package to to use for gene mapping: "gorth" (slower but more reference species) or "homologene" (faster but fewer reference species).

drop_nonorths
Drop genes that don't have an ortholog in the output_species.

one_to_one_only
Drop genes that don't have a 1:1 mapping between input species and human (i.e. drop genes with multiple human orthologs).

round_digits
Number of digits to round to when printing percentages.

return_report
Return just the ortholog mapping between two species (FALSE) or return both the ortholog mapping as well a dataframe of the report statistics (TRUE).

verbose
Print messages.

Value

List of ortholog report statistics

Examples

orth.fly <- report_orthologs(target_species = "fly",
                             reference_species="human")

#> Mapping species name: fly

#> Common name mapping found for fly

#> 1 organism identified from search: Drosophila melanogaster

#> Returning all 8,438 genes from fly.

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: Homo sapiens

#> Returning all 19,129 genes from human.

#> 
#> Converting genes: fly ===> human 

#> Preparing gene_df.

#> Extracting genes from Gene.Symbol.

#> Converting fly ==> human orthologs using homologene

#> Mapping species name: fly

#> Common name mapping found for fly

#> 1 organism identified from search: Drosophila melanogaster

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: Homo sapiens

#> Checking for genes without orthologs in human.

#> Checking for genes without 1:1 orthologs.

#> + Dropping 19 genes that have multiple input_gene per ortholog_gene.

#> + Dropping 266 genes that have multiple ortholog_gene per input_gene.

#> Adding ortholog_gene col to gene_df.

#> Genes dropped during inter-species conversion:  4,379 / 8,438 (52%)

#> 
#> 
#> =========== REPORT SUMMARY =========== 

#> 4,059 / 8,438 (48.1%) target_species genes remain after ortholog conversion.

#> 4,059 / 19,129 (21.22%) reference_species genes remain after ortholog conversion.

orth.zeb <- report_orthologs(target_species = "zebrafish", 
                             reference_species="human")

#> Mapping species name: zebrafish

#> Common name mapping found for zebrafish

#> 1 organism identified from search: Danio rerio

#> Returning all 20,897 genes from zebrafish.

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: Homo sapiens

#> Returning all 19,129 genes from human.

#> 
#> Converting genes: zebrafish ===> human 

#> Preparing gene_df.

#> Extracting genes from Gene.Symbol.

#> Converting zebrafish ==> human orthologs using homologene

#> Mapping species name: zebrafish

#> Common name mapping found for zebrafish

#> 1 organism identified from search: Danio rerio

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: Homo sapiens

#> Checking for genes without orthologs in human.

#> Checking for genes without 1:1 orthologs.

#> + Dropping 47 genes that have multiple input_gene per ortholog_gene.

#> + Dropping 2,708 genes that have multiple ortholog_gene per input_gene.

#> Adding ortholog_gene col to gene_df.

#> Genes dropped during inter-species conversion:  10,338 / 20,895 (49%)

#> 
#> 
#> =========== REPORT SUMMARY =========== 

#> 10,556 / 20,895 (50.52%) target_species genes remain after ortholog conversion.

#> 10,556 / 19,129 (55.18%) reference_species genes remain after ortholog conversion.

orth.mus <- report_orthologs(target_species = "mouse", 
                             reference_species="human") 

#> Mapping species name: mouse

#> Common name mapping found for mouse

#> 1 organism identified from search: Mus musculus

#> Returning all 21,207 genes from mouse.

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: Homo sapiens

#> Returning all 19,129 genes from human.

#> 
#> Converting genes: mouse ===> human 

#> Preparing gene_df.

#> Extracting genes from Gene.Symbol.

#> Converting mouse ==> human orthologs using homologene

#> Mapping species name: mouse

#> Common name mapping found for mouse

#> 1 organism identified from search: Mus musculus

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: Homo sapiens

#> Checking for genes without orthologs in human.

#> Checking for genes without 1:1 orthologs.

#> + Dropping 131 genes that have multiple input_gene per ortholog_gene.

#> + Dropping 498 genes that have multiple ortholog_gene per input_gene.

#> Adding ortholog_gene col to gene_df.

#> Genes dropped during inter-species conversion:  4,725 / 21,207 (22%)

#> 
#> 
#> =========== REPORT SUMMARY =========== 

#> 16,482 / 21,207 (77.72%) target_species genes remain after ortholog conversion.

#> 16,482 / 19,129 (86.16%) reference_species genes remain after ortholog conversion.