Map genes from one species to another

Currently supports 700+ species. See map_species() for a full list of available species.

convert_orthologs(
  gene_df,
  gene_col = "rownames",
  standardise_genes = FALSE,
  input_species,
  output_species = "human",
  method = c("homologene", "gorth"),
  drop_nonorths = TRUE,
  one_to_one_only = TRUE,
  genes_as_rownames = FALSE,
  as_sparse = FALSE,
  sort_rows = FALSE,
  return_dict = FALSE,
  invert_dict = FALSE,
  verbose = TRUE,
  ...
)

Arguments

gene_df
Data.frame containing the genes symbols of the input species.

gene_col
Character string indicating either the input species' gene symbols are in a column with the input species gene symbols (gene_col="<gene_col_name>") or are the row names (gene_col="rownames") or column names (gene_col="colnames").

standardise_genes
If TRUE genes will first be standardised to HGNC symbols using gprofiler2::gconvert. This will add a new column to gene_df called "input_gene_standard".

input_species
Name of the input species (e.g., "mouse","fly"). See taxa_id_dict() for a full list of available species.

output_species
Name of the output species (e.g. "human").

method
R package to to use for gene mapping: "gorth" (slower but more reference species) or "homologene" (faster but fewer reference species).

drop_nonorths
Drop genes that don't have an ortholog in the output_species.

one_to_one_only
Drop genes that don't have a 1:1 mapping between input species and human (i.e. drop genes with multiple human orthologs).

genes_as_rownames
Whether to return the data.frame with the row names set to the human orthologs.

as_sparse
If obj is a matrix, it can be converted to a sparse matrix.

sort_rows
Sort rows alphanumerically.

return_dict
Return a dictionary (named list) of gene mappings, where names are the original gene names and items are the new gene names. If return_dict=FALSE, instead returns mappings as a data.frame.

invert_dict
If return_dict=TRUE, setting invert_dict switches the names and items in the dictionary.

verbose
Print messages.

...
Additional arguments to be passed to gprofiler2::gorth() or homologene::homologene().

Value

Data.frame with the gene symbols of the input species ("Gene_orig") and converted human orthologs ("Gene").

Examples

data("exp_mouse")
gene_df <- convert_orthologs(gene_df = exp_mouse,
                             input_species="mouse",
                             genes_as_rownames=TRUE)

#> 
#> Converting genes: mouse ===> human 

#> Preparing gene_df.

#> Loading required package: Matrix

#> Extracting genes from rownames.

#> Converting mouse ==> human orthologs using homologene

#> Mapping species name: mouse

#> Common name mapping found for mouse

#> 1 organism identified from search: Mus musculus

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: Homo sapiens

#> Checking for genes without orthologs in human.

#> Checking for genes without 1:1 orthologs.

#> + Dropping 46 genes that have multiple input_gene per ortholog_gene.

#> + Dropping 56 genes that have multiple ortholog_gene per input_gene.

#> + Setting ortholog_gene col as rownames...

#> WARNING: genes_as_rownames must be TRUE when gene_df is a matrix. Setting accordingly.

#> Setting ortholog_gene to rownames.

#> Genes dropped during inter-species conversion:  2,016 / 15,259 (13%)