Map genes

Input a list of genes, transcripts, proteins, SNPs, or genomic ranges in any format (HGNC, Ensembl, RefSeq, UniProt, etc.) and return a table with standardised gene symbols (the "names" column).

map_genes(
  genes,
  species = "hsapiens",
  target = "ENSG",
  mthreshold = Inf,
  drop_na = FALSE,
  numeric_ns = "",
  verbose = TRUE
)

Arguments

genes
Gene list.

species
Species to map against.

target
target namespace.

mthreshold
maximum number of results per initial alias to show. Shows all by default.

drop_na
Drop all genes without mappings. Sets gprofiler2::gconvert(filter_na=) as well an additional round of more comprehensive NA filtering by orthogene.

numeric_ns
namespace to use for fully numeric IDs (list of available namespaces).

verbose
Print messages.

Value

Table with standardised genes.

Details

Uses gconvert. The exact contents of the output table will depend on target parameter. See ?gprofiler2::gconvert for more details.

Examples

genes <- c(
    "Klf4", "Sox2", "TSPAN12", "NM_173007", "Q8BKT6",
    "ENSMUSG00000012396", "ENSMUSG00000074637"
)
mapped_genes <- map_genes(
    genes = genes,
    species = "mouse"
)

#> Using stored `gprofiler_orgs`.

#> Mapping species name: mouse

#> Common name mapping found for mouse

#> 1 organism identified from search: mmusculus

#> 7 / 7 (100%) genes mapped.