Standardise species names

Search gprofiler database for species that match the input text string. Then translate to a standardised species ID.

map_species(
  species = NULL,
  search_cols = c("display_name", "id", "scientific_name", "taxonomy_id"),
  output_format = c("id", "display_name", "scientific_name", "taxonomy_id", "version"),
  use_genomeinfodbdata = FALSE,
  use_local = TRUE,
  verbose = TRUE
)

Arguments

species
Species query (e.g. "human", "homo sapiens", "hapiens", or 9606). If given a list, will iterate queries for each item. Set to NULL to return all species.

search_cols
Which columns to search for species substring in metadata API.

output_format
Which column to return.

use_genomeinfodbdata
Retrieve an additional 2+ million organisms from GenomeInfoDb::specData. NOTE: Not all of these organisms are available for gene/ortholog mapping.

use_local
If TRUE default, map_species uses a locally stored version of the species metadata table instead of pulling directly from the gprofiler API. Local version may not be fully up to date, but should suffice for most use cases.

verbose
Print messages.

Value

Species ID of type output_format

Examples

ids <- map_species(species = c(
    "human", 9606, "mus musculus",
    "fly", "C elegans"
))

#> Using stored `gprofiler_orgs`.

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: hsapiens

#> Mapping species name: 9606

#> 1 organism identified from search: hsapiens

#> Mapping species name: mus musculus

#> 1 organism identified from search: mmusculus

#> Mapping species name: fly

#> Common name mapping found for fly

#> 1 organism identified from search: dmelanogaster

#> Mapping species name: C elegans

#> 1 organism identified from search: celegans