Report orthologs

Identify the number of orthologous genes between two species.

report_orthologs(
  target_species = "mouse",
  reference_species = "human",
  standardise_genes = FALSE,
  method_all_genes = c("gprofiler", "homologene"),
  method_convert_orthologs = c("gprofiler", "homologene"),
  drop_nonorths = TRUE,
  non121_strategy = "drop_both_species",
  round_digits = 2,
  return_report = TRUE,
  verbose = TRUE,
  ...
)

Arguments

target_species
Target species.

reference_species
Reference species.

standardise_genes
If TRUE AND gene_output="columns", a new column "input_gene_standard" will be added to gene_df containing standardised HGNC symbols identified by gorth.

method_all_genes
R package to to use in all_genes step: "gprofiler" (slower but more species and genes) or "homologene" (faster but fewer species and genes).

method_convert_orthologs
R package to to use in convert_orthologs step: "gprofiler" (slower but more species and genes) or "homologene" (faster but fewer species and genes).

drop_nonorths
Drop genes that don't have an ortholog in the output_species.

non121_strategy
How to handle genes that don't have 1:1 mappings between input_species:output_species. Options include:

"drop_both_species" or "dbs" or 1 :
Drop genes that have duplicate mappings in either the input_species or output_species
(DEFAULT).

"drop_input_species" or "dis" or 2 :
Only drop genes that have duplicate mappings in the input_species.

"drop_output_species" or "dos" or 3 :
Only drop genes that have duplicate mappings in the output_species.

"keep_both_species" or "kbs" or 4 :
Keep all genes regardless of whether they have duplicate mappings in either species.

"keep_popular" or "kp" or 5 :
Return only the most "popular" interspecies ortholog mappings. This procedure tends to yield a greater number of returned genes but at the cost of many of them not being true biological 1:1 orthologs.

"sum","mean","median","min" or "max" :
When gene_df is a matrix and gene_output="rownames", these options will aggregate many-to-one gene mappings (input_species-to-output_species) after dropping any duplicate genes in the output_species.

round_digits
Number of digits to round to when printing percentages.

return_report
Return just the ortholog mapping between two species (FALSE) or return both the ortholog mapping as well a data.frame of the report statistics (TRUE).

verbose
Print messages.

...
Additional arguments to be passed to gorth or homologene.

NOTE: To return only the most "popular" interspecies ortholog mappings, supply mthreshold=1 here AND set method="gprofiler" above. This procedure tends to yield a greater number of returned genes but at the cost of many of them not being true biological 1:1 orthologs.

For more details, please see here.

Value

List of ortholog report statistics

Examples

orth_fly <- report_orthologs(
    target_species = "fly",
    reference_species = "human"
)

#> Retrieving all genes using: gprofiler

#> Using stored `gprofiler_orgs`.

#> Mapping species name: fly

#> Common name mapping found for fly

#> 1 organism identified from search: dmelanogaster

#> Gene table with 4592 rows retrieved.

#> Extracting genes from Gene.Symbol.

#> Dropping 100 NAs of all kinds from Gene.Symbol.

#> Returning all 4,492 genes from fly.

#> Retrieving all genes using: gprofiler

#> Using stored `gprofiler_orgs`.

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: hsapiens

#> Gene table with 60646 rows retrieved.

#> Extracting genes from Gene.Symbol.

#> Dropping 20,092 NAs of all kinds from Gene.Symbol.

#> Returning all 40,554 genes from human.

#> Preparing gene_df.

#> data.frame format detected.

#> Extracting genes from Gene.Symbol.

#> Converting fly ==> human orthologs using: gprofiler

#> Using stored `gprofiler_orgs`.

#> Mapping species name: fly

#> Common name mapping found for fly

#> 1 organism identified from search: dmelanogaster

#> Using stored `gprofiler_orgs`.

#> Mapping species name: human

#> Common name mapping found for human

#> 1 organism identified from search: hsapiens

#> Checking for genes without orthologs in human.

#> Extracting genes from input_gene.

#> Extracting genes from ortholog_gene.

#> Dropping 3,414 NAs of all kinds from ortholog_gene.

#> Checking for genes without 1:1 orthologs.

#> Dropping 1,887 genes that have multiple input_gene per ortholog_gene.

#> Dropping 748 genes that have multiple ortholog_gene per input_gene.

#> Filtering gene_df with gene_map

#> Adding input_gene col to gene_df.

#> Adding ortholog_gene col to gene_df.

#> 
#> =========== REPORT SUMMARY ===========

#> Total genes dropped after convert_orthologs :
#>    3,854 / 4,492 (86%)

#> Total genes remaining after convert_orthologs :
#>    638 / 4,492 (14%)

#> 
#> =========== REPORT SUMMARY ===========

#> 638 / 4,492 (14.2%) target_species genes remain after ortholog conversion.

#> 638 / 39,357 (1.62%) reference_species genes remain after ortholog conversion.